# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bharat Creator Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /creators:
    get:
      operationId: listCreators
      tags: [creators]
      summary: List all creators
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: language
          in: query
          schema:
            type: string
        - name: platform
          in: query
          schema:
            type: string
        - name: minFollowers
          in: query
          schema:
            type: integer
        - name: maxFollowers
          in: query
          schema:
            type: integer
        - name: state
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of creators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Creator"
    post:
      operationId: createCreator
      tags: [creators]
      summary: Create a creator profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCreatorBody"
      responses:
        "201":
          description: Created creator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creator"
  /creators/{id}:
    get:
      operationId: getCreator
      tags: [creators]
      summary: Get creator by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Creator details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creator"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCreator
      tags: [creators]
      summary: Update creator profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCreatorBody"
      responses:
        "200":
          description: Updated creator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creator"
    delete:
      operationId: deleteCreator
      tags: [creators]
      summary: Delete creator profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Creator:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        bio:
          type: string
        profileImageUrl:
          type: string
          nullable: true
        category:
          type: string
        languages:
          type: array
          items:
            type: string
        state:
          type: string
        city:
          type: string
        platforms:
          type: array
          items:
            $ref: "#/components/schemas/PlatformStat"
        totalFollowers:
          type: integer
        engagementRate:
          type: number
        avgViews:
          type: integer
        ratePerPost:
          type: number
        ratePerReel:
          type: number
        ratePerVideo:
          type: number
        tags:
          type: array
          items:
            type: string
        verified:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - username
        - bio
        - category
        - languages
        - state
        - city
        - platforms
        - totalFollowers
        - engagementRate
        - avgViews
        - ratePerPost
        - ratePerReel
        - ratePerVideo
        - tags
        - verified
        - createdAt
        - updatedAt
    PlatformStat:
      type: object
      properties:
        platform:
          type: string
          enum: [instagram, youtube, twitter, linkedin, moj, sharechat, josh, snapchat]
        handle:
          type: string
        followers:
          type: integer
        profileUrl:
          type: string
      required: [platform, handle, followers, profileUrl]
    CreateCreatorBody:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        bio:
          type: string
        profileImageUrl:
          type: string
          nullable: true
        category:
          type: string
        languages:
          type: array
          items:
            type: string
        state:
          type: string
        city:
          type: string
        platforms:
          type: array
          items:
            $ref: "#/components/schemas/PlatformStat"
        ratePerPost:
          type: number
        ratePerReel:
          type: number
        ratePerVideo:
          type: number
        tags:
          type: array
          items:
            type: string
      required:
        - name
        - username
        - bio
        - category
        - languages
        - state
        - city
        - platforms
        - ratePerPost
        - ratePerReel
        - ratePerVideo
        - tags
    UpdateCreatorBody:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        profileImageUrl:
          type: string
          nullable: true
        category:
          type: string
        languages:
          type: array
          items:
            type: string
        state:
          type: string
        city:
          type: string
        platforms:
          type: array
          items:
            $ref: "#/components/schemas/PlatformStat"
        ratePerPost:
          type: number
        ratePerReel:
          type: number
        ratePerVideo:
          type: number
        tags:
          type: array
          items:
            type: string
...